Changes: - created DTO layer - added mapstruct dependency and created mappers for DTOs - added organization, restaurant and menu context, separating data for multiple group of users/restaurants/menus - changed Integer id types for Long

+++ src/main/java/com/hackybear/hungry_scan_core/controller/cms/QrController.java
@@ -49,7 +49,7 @@ public class QrController {
 
     @PostMapping("/tables/generate-qr")
     @PreAuthorize("hasAnyRole('MANAGER', 'ADMIN')")
-    public ResponseEntity<?> generateQr(@RequestBody Integer id) {
+    public ResponseEntity<?> generateQr(@RequestBody Long id) {
         try {
             RestaurantTable restaurantTable = restaurantTableService.generateNewToken(id);
             qrService.generate(restaurantTable, "");
