+++ src/main/java/com/hackybear/hungry_scan_core/controller/restaurant/FeedbackController.java
@@ -30,7 +30,7 @@ public class FeedbackController {
 
     @PreAuthorize("hasAnyRole('MANAGER', 'ADMIN')")
     @PostMapping("/show")
-    public ResponseEntity<Map<String, Object>> show(@RequestBody Integer id) {
+    public ResponseEntity<Map<String, Object>> show(@RequestBody Long id) {
         return responseHelper.getResponseEntity(id, feedbackService::findById);
     }
